Add Quadtree and Utils
Add Quadtree, a point region quadtree. It can insert, remove and get a Point in a range randomly. The format of range is (x, y, notLessThanX, notLessThanY), which is not a rectangle. The method of getting all Points in a range is not complete and will not be finished in the foreseeable future. Add Utils for getting random numbers.