Sort empty Tiles
Modify HorizontalTilePlane and VerticalTilePlane to sort empty Tiles. Because elements in sorted containers should not be modified, add previousWidth and previousHeight to Tile to store values temporarily.  BIG UNSOLVED BUG: std::set somehow does not work with Tile. Some Tiles disappear after being inserted to a set. As a result, SortedTiles is created in replace of std::set. SortedTiles uses std::list and its time complexity is O(n). Leave the bug for the future.  See details in Trello card 21.