Add CornerSequence (not complete)
CornerSequence places Macros on Corners of CornerHorizontal/VerticalTilePlanes. It tries another Corner if the given Corner is invalid. It manages Corners with a Quadtree. It stops placing when no valid Corner is available.  Fix lots of bugs in Corner, Quadtree, Horizontal/VerticalTilePlane and CornerHorizontal/VerticalTilePlane.  A bug is not yet fixed in CornerSequence, that is, placing Macros crashes frequently with unknown reason.  Future works: 1. Degrade type 2 Corners. 2. Quadtree::remove() should merge sub Quadtrees when at most one of them is not empty. 3. Operate and copy a CornerSequence. 4. Clean up debugging code.  See Trello card 17 for details.