Disable sorting and consider space shape as cost.
Two changes: 1. Do not sort Tiles and Macros for now. Do not degrade type 2 Corners. However, doing so does not improve run time significantly. 2. Consider the largest empty space's shape as a cost. Calculate the bounding box of the empty space. Divide the empty space area with the bounding box area is called density. 1 - density is the cost. This CostFunction tends to make the space into a rectangle.  See Trello card 33 for details.