Calculate routability wirelength.
Calculate wirelength, calculate Macros density map, and use the overlap between the two to indicate the routability of a Net. If a Net's bounding box overlaps more Macros, the Net is considered having less routability. Add Matrix2d to store a density map as a 2d matrix.  See Trello card 39 for details.